Refactored and functioning version.
This version's login feature works completely and so does the Chart. An update button is added to the Chart so that it directly fetches from Airtable instead of relying on browser data which is prone to errors.

Current issues:
UI needs improvements.
Import/Export functions do not work as intended by Godfrey.